Add stage_hashstate header
