+++ node/silkworm/stagedsync/stage_hashstate.cpp
@@ -14,14 +14,13 @@
    limitations under the License.
 */
 
+#include "stage_hashstate.hpp"
+
 #include <silkworm/common/endian.hpp>
 #include <silkworm/common/log.hpp>
 #include <silkworm/db/access_layer.hpp>
-#include <silkworm/db/stages.hpp>
 #include <silkworm/db/util.hpp>
 #include <silkworm/etl/collector.hpp>
-
-#include "stagedsync.hpp"
 
 namespace silkworm::stagedsync {
