+++ node/silkworm/stagedsync/stagedsync.hpp
@@ -29,66 +29,6 @@
 
 namespace silkworm::stagedsync {
 
-class HashState final : public IStage {
-  public:
-    explicit HashState(NodeSettings* node_settings)
-        : IStage(db::stages::kHashStateKey, node_settings),
-          collector_(std::make_unique<etl::Collector>(node_settings)){};
-    ~HashState() override = default;
-    StageResult forward(db::RWTxn& txn) final;
-    StageResult unwind(db::RWTxn& txn, BlockNum to) final;
-    StageResult prune(db::RWTxn& txn) final;
-    std::vector<std::string> get_log_progress() final;
-
-  private:
-    //! \brief Store already processed addresses to avoid rehashing and multiple lookups
-    //! \struct Address -> Address Hash -> Value
-    using ChangedAddresses = absl::btree_map<evmc::address, std::pair<evmc::bytes32, Bytes>>;
-
-    //! \brief Transforms PlainState into HashedAccounts and HashedStorage respectively in one single read pass over
-    //! PlainState \remarks To be used only if this is very first time HashState stage runs forward (i.e. forwarding
-    //! from 0)
-    StageResult hash_from_plainstate(db::RWTxn& txn);
-
-    //! \brief Transforms PlainCodeHash into HashedCodeHash in one single read pass over PlainCodeHash
-    //! \remarks To be used only if this is very first time HashState stage runs forward (i.e. forwarding from 0)
-    StageResult hash_from_plaincode(db::RWTxn& txn);
-
-    //! \brief Detects account changes from AccountChangeSet and hashes the changed keys
-    //! \remarks Though it could be used for initial sync only is way slower and builds an index of changed accounts.
-    StageResult hash_from_account_changeset(db::RWTxn& txn, BlockNum previous_progress, BlockNum to);
-
-    //! \brief Detects storage changes from StorageChangeSet and hashes the changed keys
-    //! \remarks Though it could be used for initial sync only is way slower and builds an index of changed storage
-    //! locations.
-    StageResult hash_from_storage_changeset(db::RWTxn& txn, BlockNum previous_progress, BlockNum to);
-
-    //! \brief Detects account changes from AccountChangeSet and reverts hashed states
-    StageResult unwind_from_account_changeset(db::RWTxn& txn, BlockNum previous_progress, BlockNum to);
-
-    //! \brief Detects storage changes from StorageChangeSet and reverts hashed states
-    StageResult unwind_from_storage_changeset(db::RWTxn& txn, BlockNum previous_progress, BlockNum to);
-
-    //! \brief Writes to db the changes collected from account changeset scan either in forward or unwind mode
-    StageResult write_changes_from_changed_addresses(db::RWTxn& txn, const ChangedAddresses& changed_addresses);
-
-    //! \brief Writes to db the changes collected from storage changeset scan either in forward or unwind mode
-    StageResult write_changes_from_changed_storage(db::RWTxn& txn, db::StorageChanges& storage_changes,
-                                                   const absl::btree_map<evmc::address, evmc::bytes32>& hashed_addresses);
-
-    //! \brief Resets all fields related to log progress tracking
-    void reset_log_progress();
-
-    // Logger info
-    std::mutex log_mtx_{};                       // Guards async logging
-    std::atomic_bool incremental_{false};        // Whether operation is incremental
-    std::atomic_bool loading_{false};            // Whether we're in ETL loading phase
-    std::string current_source_;                 // Current source of data
-    std::string current_target_;                 // Current target of transformed data
-    std::string current_key_;                    // Actual processing key
-    std::unique_ptr<etl::Collector> collector_;  // Collector (used only in !incremental_)
-};
-
 typedef StageResult (*StageFunc)(db::RWTxn&, const std::filesystem::path& etl_path, uint64_t prune_from);
 typedef StageResult (*UnwindFunc)(db::RWTxn&, const std::filesystem::path& etl_path, uint64_t unwind_to);
 typedef StageResult (*PruneFunc)(db::RWTxn&, const std::filesystem::path& etl_path, uint64_t prune_from);
