+++ node/silkworm/stagedsync/sync_loop.cpp
@@ -19,9 +19,9 @@
 #include <boost/format.hpp>
 
 #include <silkworm/stagedsync/stage_blockhashes.hpp>
-#include <silkworm/stagedsync/stage_senders.hpp>
 #include <silkworm/stagedsync/stage_execution.hpp>
-#include <silkworm/stagedsync/stagedsync.hpp>
+#include <silkworm/stagedsync/stage_hashstate.hpp>
+#include <silkworm/stagedsync/stage_senders.hpp>
 
 namespace silkworm::stagedsync {
